Implemented a rudimentary debugger
It is activated with the -b <bp addr> command-line option.
It supports two commands, next (n) and continue (c).

This commit also ended up getting intertwined with some timing changes.
The timing now runs at full speed and ignores Chip8::freq.

This is because the amount of time to execute all pending cycles was greater
than the time available, so I ended up with a positive feedback loop in which
each frame got longer because the frame before it had so much work to do.  In
the future I will revisit this; for now I am going to hope that roms are using
the dt register to time themselves instead of relying on clock speed.