Bind the numpad keys in addition to the QWERTY keys
This makes a lot of games easier to play, but there are a few (like Blinky) that
have really awkward control schemes.  I think this is because some of them were
intended for machines with vastly different controls (like that one graphing
calculator).

There's still no interface for customizable keybinds.