inst_add_reg_reg now properly clears Vf if there was not a carry.
This fixes the sierpinski triangle demo.  There are a couple of pixels in the
lower corners which might be missing, but I don't know because I don't have
another interpreter to test on handy.  At any rate, it used to be completely
broken and now it is either wholly or mostly functional.